+++ src/page/JokesPage.vue
@@ -1,10 +1,13 @@
 <script setup lang="ts">
 import AppHeader from '@/components/organisms/AppHeader.vue';
+import AppToolBar from '@/components/organisms/AppToolBar.vue';
+
 </script>
 
 <template>
     <header>
         <AppHeader />
+        <AppToolBar />
     </header>
 </template>